Added check to ensure sync between GUI and FK.
Checking that the FinalKey responds correctly
to the triggered action. I've never seen it act wrongly,
but this will give some certainty that it's not in fact
asking to confirm another action than what the GUI requested.
This will provide very minimal protection on a compromised system,
since an intelligent attacker would easily be able to spoof the
communication, but that does not matter, since on a compromised system
an attacker may simply choose to replace the entire GUI with a hacked version.

This is for ensuring consistency and try to catch malfunctions on non-compromised
systems and should not be considered a security feature.